service
Former-commit-id: a854e8b007b31c1e075beb8c26c5c00683ce645a
Former-commit-id: 035c6c2896eb06caa18547aeabe8b0d5d4fb94ee

+++ core/service/thrift/parser/parser.go
@@ -68,7 +68,7 @@ func MemberDto(src *member.Member) *member_service.SMember {
 		Exp:            int64(src.Exp),
 		Level:          int32(src.Level),
 		PremiumUser:    int32(src.PremiumUser),
-		PremiumExpires: int64(src.PremiumExpires),
+		PremiumExpires: src.PremiumExpires,
 		InviteCode:     src.InviteCode,
 		RegIp:          src.RegIp,
 		RegFrom:        src.RegFrom,
